make gRPC + postgreSQL webservice prototype

+++ calendar/migrations/00001_create_event_table.sql
@@ -6,14 +6,13 @@ CREATE TABLE public."Event"
     id                uuid                     NOT NULL DEFAULT uuid_generate_v4() PRIMARY KEY,
     short_description CHARACTER VARYING(100)   NOT NULL,
     date              TIMESTAMP WITH TIME ZONE NOT NULL,
-    duration          INT                      NOT NULL,
+    duration          BIGINT                   NOT NULL,
     full_description  TEXT                     NULL,
-    remind_before     INT                      NULL,
+    remind_before     BIGINT                   NULL,
     user_id           uuid                     NOT NULL
 );
 
 CREATE INDEX user_idx ON public."Event" (user_id);
-
 
 ALTER TABLE public."Event"
     OWNER to postgres;
